fix: 감정 리스트를 constants로 분리
