+++ src/contents/emotionList.js
@@ -1,0 +1,9 @@
+const emotionList = [
+  { emotionId: 1, emotionName: "완전 좋음" },
+  { emotionId: 2, emotionName: "좋음" },
+  { emotionId: 3, emotionName: "그럭저럭" },
+  { emotionId: 4, emotionName: "나쁨" },
+  { emotionId: 5, emotionName: "끔찍함" },
+];
+
+export default emotionList;
